+++ src/tasks/setupMongoDb.ts
@@ -41,7 +41,7 @@ const connectDB = async () => {
 export default connectDB;
 `;
 
-    writeFile(configDbPath, configDbContent.trim(), "utf-8")
+    await writeFile(configDbPath, configDbContent.trim(), "utf-8");
 
   }